+++ runtime/src/mem.rs
@@ -693,3 +693,232 @@ pub fn seccomp_spec() -> crate::seccomp::Spec {
         ..crate::seccomp::Spec::default()
     }
 }
+
+#[cfg(unix)]
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use rand::{rngs::OsRng, Rng};
+
+    #[cfg(target_os = "linux")]
+    const MEM_ACCESS_ERR: crate::Error = crate::Error::ZoneError(crate::zone::Error::Signal { signo: libc::SIGSEGV });
+    #[cfg(target_os = "macos")]
+    const MEM_ACCESS_ERR: crate::Error = crate::Error::ZoneError(crate::zone::Error::Signal { signo: libc::SIGBUS });
+
+    #[cfg(not(feature = "stdalloc"))]
+    #[global_allocator]
+    static ALLOC: GuardedAllocator = GuardedAllocator::new();
+
+    #[cfg(not(feature = "stdalloc"))]
+    fn with_guarded_allocator<A, F: FnOnce() -> A>(f: F) -> A {
+        f()
+    }
+
+    #[cfg(feature = "stdalloc")]
+    fn with_guarded_allocator<A, F: FnOnce() -> A>(f: F) -> A {
+        unsafe { stdalloc::guarded() };
+        let a = f();
+        unsafe { stdalloc::std() };
+        a
+    }
+
+    fn page_size_exponent() -> u32 {
+        let mut p = 1;
+        let mut k = 0;
+        while p != page_size() {
+            p *= 2;
+            k += 1;
+        }
+        k as u32
+    }
+
+    fn fresh_non_zero_size(bound: usize) -> usize {
+        let mut n = 0;
+        while n == 0 {
+            n = OsRng.gen::<usize>() % bound;
+        }
+        n
+    }
+
+    fn fresh_layout() -> Layout {
+        let n = fresh_non_zero_size(3 * page_size());
+        let a = 2usize.pow(OsRng.gen::<u32>() % page_size_exponent() + 3);
+        Layout::from_size_align(n, a).unwrap()
+    }
+
+    fn do_test_write(p: *mut u8, n: usize) {
+        let bs = unsafe { core::slice::from_raw_parts_mut(p, n) };
+        for b in bs.iter() {
+            assert_eq!(*b, 0u8);
+        }
+
+        OsRng.fill(bs);
+    }
+
+    fn do_sized_alloc_test(n: usize) -> crate::Result<()> {
+        let a = GuardedAllocation::unaligned(n)?;
+
+        do_test_write(a.data(), n);
+
+        a.free()?;
+
+        Ok(())
+    }
+
+    #[test]
+    fn allocate_whole_page() -> crate::Result<()> {
+        do_sized_alloc_test(page_size())
+    }
+
+    #[test]
+    fn allocate_less_than_a_whole_page() -> crate::Result<()> {
+        do_sized_alloc_test(1)
+    }
+
+    #[test]
+    fn allocate_little_more_than_a_whole_page() -> crate::Result<()> {
+        do_sized_alloc_test(page_size() + 1)
+    }
+
+    #[test]
+    fn allocate_random_sizes() -> crate::Result<()> {
+        for _ in 1..10 {
+            do_sized_alloc_test(fresh_non_zero_size(1024 * 1024))?
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn alignment() -> crate::Result<()> {
+        for _ in 1..100 {
+            let l = fresh_layout();
+            let a = GuardedAllocation::aligned(l)?;
+            assert_eq!((a.data() as usize) % l.align(), 0);
+            do_test_write(a.data(), l.size());
+            a.free()?;
+        }
+
+        Ok(())
+    }
+
+    #[test]
+    fn zero_allocation() -> crate::Result<()> {
+        assert_eq!(GuardedAllocation::unaligned(0), Err(Error::ZeroAllocation.into()),);
+        Ok(())
+    }
+
+    #[test]
+    fn guard_pages_pre_read() -> crate::Result<()> {
+        let l = fresh_layout();
+        let a = GuardedAllocation::aligned(l)?;
+
+        assert_eq!(
+            crate::zone::fork(|| {
+                for i in 0..page_size() {
+                    unsafe {
+                        assert_eq!(0u8, core::ptr::read_unaligned(a.data().offset(-(i as isize))));
+                    }
+                }
+            }),
+            Err(MEM_ACCESS_ERR)
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn guard_pages_pre_write() -> crate::Result<()> {
+        let l = fresh_layout();
+        let a = GuardedAllocation::aligned(l)?;
+
+        assert_eq!(
+            crate::zone::fork(|| {
+                for i in 0..page_size() {
+                    unsafe {
+                        core::ptr::write_unaligned(a.data().offset(-(i as isize)), OsRng.gen());
+                    }
+                }
+            }),
+            Err(MEM_ACCESS_ERR)
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn guard_pages_post_read() -> crate::Result<()> {
+        let l = fresh_layout();
+        let a = GuardedAllocation::aligned(l)?;
+
+        assert_eq!(
+            crate::zone::fork(|| {
+                for i in 0..page_size() {
+                    unsafe {
+                        assert_eq!(0u8, core::ptr::read_unaligned(a.data().add(l.size() + i)));
+                    };
+                }
+            }),
+            Err(MEM_ACCESS_ERR)
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn guard_pages_post_write() -> crate::Result<()> {
+        let l = fresh_layout();
+        let a = GuardedAllocation::aligned(l)?;
+
+        assert_eq!(
+            crate::zone::fork(|| {
+                for i in 0..page_size() {
+                    unsafe {
+                        core::ptr::write_unaligned(a.data().add(l.size() + i), OsRng.gen());
+                    }
+                }
+            }),
+            Err(MEM_ACCESS_ERR)
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn vectors() -> crate::Result<()> {
+        with_guarded_allocator(|| {
+            extern crate alloc;
+            use alloc::vec::Vec;
+
+            let mut bs: Vec<u8> = Vec::with_capacity(10);
+            for _ in 1..100 {
+                bs.push(OsRng.gen());
+            }
+
+            Ok(())
+        })
+    }
+
+    // TODO: unify these apis, maybe a dedicated zone::Spec?
+    #[test]
+    #[cfg(target_os = "linux")]
+    fn inside_zone_linux() -> crate::Result<()> {
+        let l = fresh_layout();
+        crate::zone::fork(|| {
+            seccomp_spec().with_getrandom().apply().unwrap();
+            let a = GuardedAllocation::aligned(l).unwrap();
+            do_test_write(a.data(), l.size());
+            a.free().unwrap();
+        })
+    }
+
+    #[test]
+    #[cfg(target_os = "macos")]
+    fn inside_zone_macos() -> crate::Result<()> {
+        let l = fresh_layout();
+        crate::zone::fork(|| {
+            let a = GuardedAllocation::aligned(l).unwrap();
+            do_test_write(a.data(), l.size());
+            a.free().unwrap();
+        })
+    }
+}
